+++ test/integration/cross_serialization/ClientSerialization.cpp
@@ -10,7 +10,7 @@ std::vector<uint8_t> ClientSerialization::create_client_payload()
     std::vector<uint8_t> buffer(BUFFER_LENGTH, 0x00);
 
     MicroBuffer micro_buffer;
-    init_micro_buffer(&micro_buffer, &buffer.front(), buffer.capacity());
+    init_micro_buffer(&micro_buffer, &buffer.front(), (uint32_t)buffer.capacity());
 
     CREATE_CLIENT_Payload payload;
     payload.base.request_id = RequestId{0x01, 0x23};
@@ -35,7 +35,7 @@ std::vector<uint8_t> ClientSerialization::create_payload()
     std::vector<uint8_t> buffer(BUFFER_LENGTH, 0x00);
 
     MicroBuffer micro_buffer;
-    init_micro_buffer(&micro_buffer, &buffer.front(), buffer.capacity());
+    init_micro_buffer(&micro_buffer, &buffer.front(), (uint32_t)buffer.capacity());
 
     CREATE_Payload payload;
     payload.base.request_id = RequestId{0x01, 0x23};
@@ -58,7 +58,7 @@ std::vector<uint8_t> ClientSerialization::get_info_payload()
     std::vector<uint8_t> buffer(BUFFER_LENGTH, 0x00);
 
     MicroBuffer micro_buffer;
-    init_micro_buffer(&micro_buffer, &buffer.front(), buffer.capacity());
+    init_micro_buffer(&micro_buffer, &buffer.front(), (uint32_t)buffer.capacity());
 
     GET_INFO_Payload payload;
     payload.base.request_id = RequestId{0x01, 0x23};
@@ -76,7 +76,7 @@ std::vector<uint8_t> ClientSerialization::delete_payload()
     std::vector<uint8_t> buffer(BUFFER_LENGTH, 0x00);
 
     MicroBuffer micro_buffer;
-    init_micro_buffer(&micro_buffer, &buffer.front(), buffer.capacity());
+    init_micro_buffer(&micro_buffer, &buffer.front(), (uint32_t)buffer.capacity());
 
     DELETE_Payload payload;
     payload.base.request_id = RequestId{0x01, 0x23};
@@ -99,7 +99,7 @@ std::vector<uint8_t> ClientSerialization::status_payload()
     std::vector<uint8_t> buffer(BUFFER_LENGTH, 0x00);
 
     MicroBuffer micro_buffer;
-    init_micro_buffer(&micro_buffer, &buffer.front(), buffer.capacity());
+    init_micro_buffer(&micro_buffer, &buffer.front(), (uint32_t)buffer.capacity());
 
     STATUS_Payload payload;
     payload.base.related_request.request_id = RequestId{0x01, 0x23};
@@ -124,7 +124,7 @@ std::vector<uint8_t> ClientSerialization::read_data_payload()
     std::vector<uint8_t> buffer(BUFFER_LENGTH, 0x00);
 
     MicroBuffer micro_buffer;
-    init_micro_buffer(&micro_buffer, &buffer.front(), buffer.capacity());
+    init_micro_buffer(&micro_buffer, &buffer.front(), (uint32_t)buffer.capacity());
 
     READ_DATA_Payload payload;
     payload.base.request_id = RequestId{0x01, 0x23};
@@ -150,7 +150,7 @@ std::vector<uint8_t> ClientSerialization::write_data_payload_data()
     std::vector<uint8_t> buffer(BUFFER_LENGTH, 0x00);
 
     MicroBuffer micro_buffer;
-    init_micro_buffer(&micro_buffer, &buffer.front(), buffer.capacity());
+    init_micro_buffer(&micro_buffer, &buffer.front(), (uint32_t)buffer.capacity());
 
     WRITE_DATA_Payload_Data payload;
     payload.base.request_id = RequestId{0x01, 0x23};
@@ -193,7 +193,7 @@ std::vector<uint8_t> ClientSerialization::data_payload_data()
     std::vector<uint8_t> buffer(BUFFER_LENGTH, 0x00);
 
     MicroBuffer micro_buffer;
-    init_micro_buffer(&micro_buffer, &buffer.front(), buffer.capacity());
+    init_micro_buffer(&micro_buffer, &buffer.front(), (uint32_t)buffer.capacity());
 
     DATA_Payload_Data payload;
     payload.base.request_id = RequestId{0x01, 0x23};
@@ -236,7 +236,7 @@ std::vector<uint8_t> ClientSerialization::acknack_payload()
     std::vector<uint8_t> buffer(BUFFER_LENGTH, 0x00);
 
     MicroBuffer micro_buffer;
-    init_micro_buffer(&micro_buffer, &buffer.front(), buffer.capacity());
+    init_micro_buffer(&micro_buffer, &buffer.front(), (uint32_t)buffer.capacity());
 
     ACKNACK_Payload payload;
     payload.first_unacked_seq_num = (uint16_t)0x0123;
@@ -254,7 +254,7 @@ std::vector<uint8_t> ClientSerialization::heartbeat_payload()
     std::vector<uint8_t> buffer(BUFFER_LENGTH, 0x00);
 
     MicroBuffer micro_buffer;
-    init_micro_buffer(&micro_buffer, &buffer.front(), buffer.capacity());
+    init_micro_buffer(&micro_buffer, &buffer.front(), (uint32_t)buffer.capacity());
 
     HEARTBEAT_Payload payload;
     payload.first_unacked_seq_nr = (uint16_t)0x0123;
